Database tables columns are indexed for performance

+++ database/migrations/2019_01_10_002911_create_shopify_shops_table.php
@@ -19,7 +19,7 @@ class CreateShopifyShopsTable extends Migration
         Schema::create('shopify_shops', function (Blueprint $table) {
             $table->increments('id');
             $table->mediumText('access_token');
-            $table->unsignedInteger('store_id');
+            $table->bigInteger('store_id');
             $table->string('name', 255)->nullable();
             $table->string('domain', 255)->nullable();
             $table->string('myshopify_domain', 255)->nullable();
